+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Since last release
 
 **Added:**
 * Added package parameter to storage (#603, #612)
+* Added package parameter to source (#613)
 
 **Changed:**